requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
openai>=1.40.0
